src/main/java/io/kirill/customer/CustomerService.java
package io.kirill.customer;

import io.kirill.customer.exceptions.CustomerNotFoundException;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class CustomerService {
    private final CustomerRepository customerRepository;

    public Customer create(String email) {
        var customer = new Customer(email);
        return customerRepository.save(customer);
    }

    public Customer get(Long id) {
        return customerRepository.findById(id).orElseThrow(() -> new CustomerNotFoundException(id));
    }
}
